Added curl function and filter function
